Update further-operations-on-delegators.md
Since the selection of an action starts always with the selection of the category first, I suggest to add it whenever it is used. Alternatively, mention the category selection somewhere at the top, e.g. Line #3 so it would not be needed to repeat again for any action covered below / the rest of site.